+++ src/StateSmith/Runner/DiscardingConsolePrinter.cs
@@ -1,0 +1,17 @@
+﻿namespace StateSmith.Runner;
+
+// we may want to capture this info at some point
+public class DiscardingConsolePrinter : IConsolePrinter
+{
+    public void WriteErrorLine(string message)
+    {
+    }
+
+    public void WriteLine()
+    {
+    }
+
+    public void WriteLine(string message)
+    {
+    }
+}